Refactoring to prepare for the TUF updates feature (#196)
* extract TUF root client routines into a separate file;
* rename TUF related routines to hold Tuf prefix or suffix where applicable;
* make private all TUF related functions (which are not re-used outside the subcommands/keys package;
* use canonical MarshalIndent to output user-friendly JSON data in all commands;
* some other minor non-functional code cleanups.
