+++ subcommands/keys/tuf_copy_targets.go
@@ -42,12 +42,12 @@ func doCopyTargets(cmd *cobra.Command, args []string) {
 
 	targetsCreds, err := createTargetsCreds(factory, *root, creds)
 	subcommands.DieNotNil(err)
-	SaveCreds(args[1], targetsCreds)
+	saveTufCreds(args[1], targetsCreds)
 }
 
 func createTargetsCreds(factory string, root client.AtsTufRoot, creds OfflineCreds) (OfflineCreds, error) {
 	targets := make(OfflineCreds)
-	onlinePub, err := api.GetFoundriesTargetsKey(factory)
+	onlinePub, err := api.TufTargetsOnlineKey(factory)
 	subcommands.DieNotNil(err)
 	for _, keyid := range root.Signed.Roles["targets"].KeyIDs {
 		pubkey := root.Signed.Keys[keyid].KeyValue.Public
