+++ subcommands/targets/edit.go
@@ -35,7 +35,7 @@ func doEdit(cmd *cobra.Command, args []string) {
 	// Get raw json
 	targets, err := api.TargetsList(factory)
 	subcommands.DieNotNil(err)
-	orig, err := json.MarshalIndent(targets, "", "  ")
+	orig, err := subcommands.MarshalIndent(targets, "", "  ")
 	subcommands.DieNotNil(err)
 
 	// Create temp file to edit with
